requirements.txt
aiohttp>=3.9.0
anyio>=4.4.0 # not directly required, pinned by Snyk to avoid a vulnerability
beautifulsoup4
click>=8.1.3
debugpy
docstring-parser>=0.16
fsspec>=2021.09.0
httpx
ijson>=3.2.0
jsonlines>=3.0.0
jsonpatch>=1.32
jsonschema>3.1.1
mmh3>3.1.0
nest_asyncio
numpy
platformdirs>=2.3.0
psutil
pydantic>=2
python-dotenv>=0.16.0
pyyaml
rich>=13.3.3
s3fs>=2023
semver>=3.0.0
shortuuid
tenacity
textual>=0.86.2
typing_extensions>=4.9.0
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability

# our additional dependencies
fastapi >= 0.115.0
uvicorn >= 0.34.0
